+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Revision history for byte-order
 
-## 0.1.0.0 -- YYYY-mm-dd
+## 0.1.1.0 -- YYYY-mm-dd
 
-* First version. Released on an unsuspecting world.
+* Add `PrimUnaligned` instance for `Fixed`.
+* Add modules for more convenient interface to reading and writing
+  fixed-endianness elements to byte arrays:
+  `Data.Primitive.ByteArray.LittleEndian` and
+  `Data.Primitive.ByteArray.BigEndian`.
+
+## 0.1.0.0 -- 2019-05-29
+
+* Initial release.